requirements.txt
Flask
mysql-connector-python
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
bcrypt